move core encryption logic to utils
