+++ packages/records/package.json
@@ -16,7 +16,6 @@
         "directory": "packages/records"
     },
     "dependencies": {
-        "@nangohq/shared": "file:../shared",
         "@nangohq/utils": "file:../utils",
         "dayjs": "1.11.10",
         "knex": "3.1.0",
