+++ packages/utils/lib/index.ts
@@ -5,3 +5,4 @@ export * from './errors.js';
 export * from './logger.js';
 export * from './id.js';
 export * from './result.js';
+export * from './encryption.js';
